+++ changelog.md
@@ -8,6 +8,7 @@ See also the commit log at github: https://github.com/Arduino-IRremote/Arduino-I
 - 2. phase renamed *.cpp.h to .hpp.
 - No deprecation print for ATtinies.
 - Renamed ac_LG.cpp to ac_LG.hpp.
+- Maintained MagiQuest by E. Stuart Hicks.
 
 ## 3.4.0
 - Added LG2 protocol.